services/users/requirements.txt
aniso8601==10.0.1
black==24.10.0
blinker==1.9.0
click==8.2.1
coverage==7.6.1
flake8==7.1.1
pytest==8.4.2
pytest-cov==6.0.0
Flask==3.1.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
flask-restx==1.3.0
gunicorn==19.9.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2025.2
six==1.17.0
Werkzeug==3.1.3
PyYAML==6.0.1
psutil==5.9.6
